pyproject.toml
[project]
name = "betfinder"
version = "0.1.0"
description = "BetFinder is a desktop/hosted application that helps you find the best odds for your bets."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.1",
    "apscheduler>=3.11.2",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
    "asyncpg>=0.30.0",
]

[dependency-groups]
dev = [
    "pystray>=0.19.5",
    "pillow>=10.0.0",
    "pyinstaller>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
dev = "app.main:run_dev"
prod = "app.main:run_prod"
build = "build_tray:build_app"
installer = "build_installer:build_installer"

[tool.hatch.build.targets.wheel]
packages = ["app"]
